+++ python/ray/data/_internal/plan.py
@@ -8,12 +8,6 @@ from ray._private.internal_api import get_memory_info_reply, get_state_from_addr
 from ray.data._internal.execution.interfaces import RefBundle
 from ray.data._internal.logical.interfaces import SourceOperator
 from ray.data._internal.logical.interfaces.logical_plan import LogicalPlan
-from ray.data._internal.logical.interfaces.operator import Operator
-from ray.data._internal.logical.optimizers import (
-    LogicalOptimizer,
-    PhysicalOptimizer,
-)
-from ray.data._internal.planner import create_planner
 from ray.data._internal.stats import DatasetStats
 from ray.data.block import _take_first_non_empty_schema
 from ray.data.context import DataContext
@@ -24,7 +18,6 @@ if TYPE_CHECKING:
     from ray.data._internal.execution.streaming_executor import (
         StreamingExecutor,
     )
-    from ray.data.dataset import Dataset
 
 
 # Scheduling strategy can be inherited from prev operator if not specified.
@@ -110,215 +103,6 @@ class ExecutionPlan:
             f"snapshot_operator={self._cache._operator}"
             f")"
         )
-
-    def explain(self) -> str:
-        """Return a string representation of the logical and physical plan."""
-        sections = []
-
-        def _add_section(title, plan):
-            plan_str, _ = self.generate_plan_string(plan.dag, show_op_repr=True)
-            banner = f"\n-------- {title} --------\n"
-            sections.append(f"{banner}{plan_str}")
-
-        # 1. Logical Plan
-        logical_plan = self._logical_plan
-        _add_section("Logical Plan", logical_plan)
-
-        # 2. Optimized Logical Plan
-        optimized_logical = LogicalOptimizer().optimize(logical_plan)
-        _add_section("Logical Plan (Optimized)", optimized_logical)
-
-        # 3. Physical Plan
-        physical_plan, _ = create_planner().plan(optimized_logical)
-        _add_section("Physical Plan", physical_plan)
-
-        # 4. Optimized Physical Plan
-        optimized_physical = PhysicalOptimizer().optimize(physical_plan)
-        _add_section("Physical Plan (Optimized)", optimized_physical)
-
-        return "".join(sections)
-
-    @staticmethod
-    def generate_plan_string(
-        op: Operator,
-        curr_str: str = "",
-        depth: int = 0,
-        including_source: bool = True,
-        show_op_repr: bool = False,
-    ):
-        """Traverse (DFS) the Plan DAG and
-        return a string representation of the operators."""
-        if not including_source and isinstance(op, SourceOperator):
-            return curr_str, depth
-
-        curr_max_depth = depth
-
-        # For logical plan, only show the operator name like "Aggregate".
-        # But for physical plan, show the operator class name as well like "AllToAllOperator[Aggregate]".
-        op_str = repr(op) if show_op_repr else op.name
-
-        if depth == 0:
-            curr_str += f"{op_str}\n"
-        else:
-            trailing_space = " " * ((depth - 1) * 3)
-            curr_str += f"{trailing_space}+- {op_str}\n"
-
-        for input in op.input_dependencies:
-            curr_str, input_max_depth = ExecutionPlan.generate_plan_string(
-                input, curr_str, depth + 1, including_source, show_op_repr
-            )
-            curr_max_depth = max(curr_max_depth, input_max_depth)
-        return curr_str, curr_max_depth
-
-    def get_plan_as_string(self, dataset: "Dataset") -> str:
-        """Create a cosmetic string representation of this execution plan.
-
-        Returns:
-            The string representation of this execution plan.
-        """
-        # NOTE: this is used for Dataset.__repr__ to give a user-facing string
-        # representation. Ideally ExecutionPlan.__repr__ should be replaced with this
-        # method as well.
-
-        from ray.data.dataset import MaterializedDataset
-
-        dataset_cls = dataset.__class__
-
-        # Do not force execution for schema, as this method is expected to be very
-        # cheap.
-        plan_str = ""
-        plan_max_depth = 0
-
-        if not self.has_computed_output():
-            # using dataset as source here, so don't generate
-            # source operator in generate_plan_string
-            plan_str, plan_max_depth = self.generate_plan_string(
-                self._logical_plan.dag, including_source=False
-            )
-
-        schema = dataset._base_schema(fetch_if_missing=False)
-        count = self._cache.get_num_rows(self._logical_plan.dag)
-
-        if schema is None or count is None:
-            has_n_ary_operator = False
-            dag = self._logical_plan.dag
-
-            while not isinstance(dag, SourceOperator):
-                if len(dag.input_dependencies) > 1:
-                    has_n_ary_operator = True
-                    break
-
-                dag = dag.input_dependencies[0]
-
-            # TODO(@bveeramani): Handle schemas for n-ary operators like `Union`.
-            if not has_n_ary_operator:
-                assert isinstance(dag, SourceOperator), dag
-                # We infer from logical plan's dag directly as we know that
-                # we don't have any cached values, so inferring is the only
-                # option left.
-                if schema is None:
-                    schema = dag.infer_schema()
-                if count is None:
-                    count = dag.infer_metadata().num_rows
-
-        if schema is None:
-            schema_str = "Unknown schema"
-        elif isinstance(schema, type):
-            schema_str = str(schema)
-        else:
-            schema_str = []
-            for n, t in zip(schema.names, schema.types):
-                if hasattr(t, "__name__"):
-                    t = t.__name__
-                schema_str.append(f"{n}: {t}")
-            schema_str = ", ".join(schema_str)
-            schema_str = "{" + schema_str + "}"
-
-        if count is None:
-            count = "?"
-
-        num_blocks = None
-        if dataset_cls == MaterializedDataset:
-            num_blocks = self.initial_num_blocks()
-            assert num_blocks is not None
-
-        name_str = (
-            "name={}, ".format(self._dataset_name)
-            if self._dataset_name is not None
-            else ""
-        )
-        num_blocks_str = f"num_blocks={num_blocks}, " if num_blocks else ""
-
-        dataset_str = "{}({}{}num_rows={}, schema={})".format(
-            dataset_cls.__name__,
-            name_str,
-            num_blocks_str,
-            count,
-            schema_str,
-        )
-
-        # If the resulting string representation fits in one line, use it directly.
-        SCHEMA_LINE_CHAR_LIMIT = 80
-        MIN_FIELD_LENGTH = 10
-        INDENT_STR = " " * 3
-        trailing_space = INDENT_STR * plan_max_depth
-
-        if len(dataset_str) > SCHEMA_LINE_CHAR_LIMIT:
-            # If the resulting string representation exceeds the line char limit,
-            # first try breaking up each `Dataset` parameter into its own line
-            # and check if each line fits within the line limit. We check the
-            # `schema` param's length, since this is likely the longest string.
-            schema_str_on_new_line = f"{trailing_space}{INDENT_STR}schema={schema_str}"
-            if len(schema_str_on_new_line) > SCHEMA_LINE_CHAR_LIMIT:
-                # If the schema cannot fit on a single line, break up each field
-                # into its own line.
-                schema_str = []
-                for n, t in zip(schema.names, schema.types):
-                    if hasattr(t, "__name__"):
-                        t = t.__name__
-                    col_str = f"{trailing_space}{INDENT_STR * 2}{n}: {t}"
-                    # If the field line exceeds the char limit, abbreviate
-                    # the field name to fit while maintaining the full type
-                    if len(col_str) > SCHEMA_LINE_CHAR_LIMIT:
-                        shortened_suffix = f"...: {str(t)}"
-                        # Show at least 10 characters of the field name, even if
-                        # we have already hit the line limit with the type.
-                        chars_left_for_col_name = max(
-                            SCHEMA_LINE_CHAR_LIMIT - len(shortened_suffix),
-                            MIN_FIELD_LENGTH,
-                        )
-                        col_str = (
-                            f"{col_str[:chars_left_for_col_name]}{shortened_suffix}"
-                        )
-                    schema_str.append(col_str)
-                schema_str = ",\n".join(schema_str)
-                schema_str = (
-                    "{\n" + schema_str + f"\n{trailing_space}{INDENT_STR}" + "}"
-                )
-            name_str = (
-                f"\n{trailing_space}{INDENT_STR}name={self._dataset_name},"
-                if self._dataset_name is not None
-                else ""
-            )
-            num_blocks_str = (
-                f"\n{trailing_space}{INDENT_STR}num_blocks={num_blocks},"
-                if num_blocks
-                else ""
-            )
-            dataset_str = (
-                f"{dataset_cls.__name__}("
-                f"{name_str}"
-                f"{num_blocks_str}"
-                f"\n{trailing_space}{INDENT_STR}num_rows={count},"
-                f"\n{trailing_space}{INDENT_STR}schema={schema_str}"
-                f"\n{trailing_space})"
-            )
-
-        if plan_max_depth == 0:
-            plan_str += dataset_str
-        else:
-            plan_str += f"{INDENT_STR * (plan_max_depth - 1)}+- {dataset_str}"
-        return plan_str
 
     def link_logical_plan(self, logical_plan: "LogicalPlan"):
         """Link the logical plan into this execution plan.
